Blacklistmodel Added
This script blocks access to <blacklist> machines. If a player gets into a blacklist, he will automatically drop him down and receive a message that he has been warned that he can not drive.

Update for the future: If he has a job as a cop, he will be able to drive special vehicles.